requirements.txt
openai==1.0.0
python-dotenv==0.19.2
requests==2.26.0
waitress==2.0.0
Flask==2.0.1
pytz==2021.3
redis==4.5.1 